refactor utils and fix prompts test
